+++ src/tracer.rs
@@ -156,22 +156,15 @@ impl<B: Binary, S: ReturnStack> Tracer<B, S> {
         match sync {
             Synchronization::Start(start) => {
                 let is_tracing = self.iter_state.is_tracing();
-                let version = self.version;
 
                 let mut initer = self.sync_init(start.address, !is_tracing, !start.branch)?;
                 if is_tracing && previous != Some(Event::Trap { thaddr: false }) {
-                    let action = match version {
-                        Version::V1 => state::SyncAction::Compare,
-                        _ => state::SyncAction::Update,
-                    };
                     initer.set_condition(state::StopCondition::Sync {
                         context: start.ctx.into(),
-                        action,
+                        action: state::SyncAction::Update,
                     });
                 } else {
-                    if version != Version::V1 {
-                        initer.set_context(start.ctx.into());
-                    }
+                    initer.set_context(start.ctx.into());
                     initer.reset_to_address()?;
                     self.iter_state = IterationState::ContextItem {
                         pc: None,
@@ -197,11 +190,8 @@ impl<B: Binary, S: ReturnStack> Tracer<B, S> {
                     initer.set_address(trap.address);
                     initer.reset_to_address()?;
                 } else {
-                    let version = self.version;
                     let mut initer = self.sync_init(trap.address, false, !trap.branch)?;
-                    if version != Version::V1 {
-                        initer.set_context(trap.ctx.into());
-                    }
+                    initer.set_context(trap.ctx.into());
                     initer.reset_to_address()?;
                 }
                 self.iter_state = IterationState::TrapItem {
@@ -214,9 +204,7 @@ impl<B: Binary, S: ReturnStack> Tracer<B, S> {
             Synchronization::Context(ctx) => {
                 let mut initer = self.state.initializer(&mut self.binary)?;
                 initer.set_stack_depth(None);
-                if self.version != Version::V1 {
-                    initer.set_context(ctx.into());
-                }
+                initer.set_context(ctx.into());
                 self.iter_state = IterationState::ContextItem {
                     pc: None,
                     context: ctx.into(),