+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file. Its format
 is based on https://keepachangelog.com/en/1.1.0/.
 
-## Unreleased
+## 0.5.2 - 2025-11-03
 
 ### Fixed